Rewritten TIFF loading to support float images, amongst other things.
This is not yet fully tested, so it is likely that there are some bugs.
Get the previous known working version with the tag:
BEFORE_IMAGE_LOAD_REWRITE_2008_08_01

The switch between compile time types and run time types has been replaced with
compile/run time iteration over a typelist. Currently, only the TIFF loading
mechanism has been written to use the automatic iteration.

+++ pnm_src/save_postscript.cxx
@@ -81,7 +81,7 @@ namespace PS
 	unsigned long int num = 0;
 	
 	for(int i=0; i <n; i++)
-		num |= buf[i] << ((3-i) * 8);
+		num |= ((unsigned long)buf[i]) << ((3-i) * 8);
 	
 	if(num == 0)
 		return "z";
